in the middle of updating the const reg duplication pass
